Refactoring: moved getSettingAsSet() method and DEFAULT_DISABLED_CATEGORIES from AuditConfig to ConfigConstants.
(cherry picked from commit d6934c67a96a4603f1a3394a7f481661d3b07d43)
